task: prevent double hits from being possible.
ISSUE: Previously, if we hit the same location 3 times the function would incement the numOfHits to 3 and we would get a "kill". However, we want to ensure that if we hit the same spot again that it doesnt count as a hit.

WORK: Changed from an enhanced for loop to a regular forloop so that we can change the ships location value that weve already guessed to -1. This way we can check if -1 exists and report a miss.